requirements.txt
flask
openai
python-dotenv
sentence-transformers

